Navbar als een component gemaakt
